+++ common/repositories/base/base_character_corpse_items_repository.h
@@ -35,17 +35,17 @@
 class BaseCharacterCorpseItemsRepository {
 public:
 	struct CharacterCorpseItems {
-		int   corpse_id;
-		int   equip_slot;
-		int   item_id;
-		int   charges;
-		int   aug_1;
-		int   aug_2;
-		int   aug_3;
-		int   aug_4;
-		int   aug_5;
-		int   aug_6;
-		int16 attuned;
+		int corpse_id;
+		int equip_slot;
+		int item_id;
+		int charges;
+		int aug_1;
+		int aug_2;
+		int aug_3;
+		int aug_4;
+		int aug_5;
+		int aug_6;
+		int attuned;
 	};
 
 	static std::string PrimaryKey()
