+++ common/repositories/base/base_faction_list_mod_repository.h
@@ -37,7 +37,7 @@ public:
 	struct FactionListMod {
 		int         id;
 		int         faction_id;
-		int16       mod;
+		int         mod;
 		std::string mod_name;
 	};
