+++ common/repositories/base/base_raid_details_repository.h
@@ -37,7 +37,7 @@ public:
 	struct RaidDetails {
 		int         raidid;
 		int         loottype;
-		int8        locked;
+		int         locked;
 		std::string motd;
 	};
